pyproject.toml
[project]
name = "factory_model"
version = "0.1.0"
description = "Una librería para autenticación y gestión de modelos LLM"
authors = [
    { name="Jorge San José", email="jsanjose@profileit.es" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.72.0",
    "azure-core>=1.33.0",
    "azure-identity>=1.21.0",
    "langchain>=0.3.23",
    "langchain_openai>=0.3.12",
    "langchain_azure_ai>=0.1.2",
    "langchain_google_genai>=2.1.2",
    "langchain_community>=0.3.21",
    "langchain_ollama>=0.3.1",
    "jinja2>=3.1.6",
    "python-decouple>=3.8",
    "pyyaml>=6.0.2"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"